matlab: Change a data type in the C code to fix 64-bit builds
Do not use "unsigned long" to store a 32-bit field.

+++ matlab/read_bfee.c
@@ -6,7 +6,7 @@
 /* The computational routine */
 void read_bfee(unsigned char *inBytes, mxArray *outCell)
 {
-	unsigned long timestamp_low = inBytes[0] + (inBytes[1] << 8) +
+	unsigned int timestamp_low = inBytes[0] + (inBytes[1] << 8) +
 		(inBytes[2] << 16) + (inBytes[3] << 24);
 	unsigned short bfee_count = inBytes[4] + (inBytes[5] << 8);
 	unsigned int Nrx = inBytes[8];
